Move CounterApp in to components
Move CounterApp in to components
